mathilo_ecommerce/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-paypal==2.0
djangorestframework==3.14.0
idna==3.6
mysqlclient==2.2.0
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
